% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{mclexplorer.run}
\alias{mclexplorer.run}
\title{Run MCLexplorer}
\usage{
mclexplorer.run()
}
\description{
Function to launch the MCLexplorer Shiny application.
}
\examples{
\dontrun{
mclexplorer.run()
}
}
\author{
Pernice Simone
}
