% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{ShinyMCL.run}
\alias{ShinyMCL.run}
\title{Run ShinyMCL}
\usage{
ShinyMCL.run()
}
\description{
function to lunch the ShinyMCL shiny application.
}
\examples{
\dontrun{
ShinyMCL.run()
}
}
\author{
Pernice Simone
}
