% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{mclexplorer.run}
\alias{mclexplorer.run}
\title{Run mclexplorer}
\usage{
mclexplorer.run()
}
\description{
function to lunch the mclexplorer shiny application.
}
\examples{
\dontrun{
mclexplorer.run()
}
}
\author{
Pernice Simone
}
